+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - TestEZ for testing
 - Functionality for using negated permission nodes in Permissions & Group classes
 - Group.SetPrecedence() used to assign a group precedence over others
+- Group.SetInheritant() used to "inherit" another groups permissions
 - Internal function isNodeNegated() to check for negated permission nodes
 ### Changed
 - Permissions.Init() to take a table as a config rather then ModuleScript